landing/content-pipeline/requirements.txt
anthropic>=0.43
Pillow>=11.0
requests>=2.32
PyJWT>=2.9
markdown>=3.7
python-frontmatter>=1.1
python-slugify>=8.0
python-dotenv>=1.0
beautifulsoup4>=4.12
